feat: добавлена Spring Security аутентификация

+++ src/main/java/ru/blog_app/blog/service/impl/PostServiceImpl.java
@@ -1,8 +1,10 @@
 package ru.blog_app.blog.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
-import ru.blog_app.blog.dto.PostDtoResponse;
+import ru.blog_app.blog.dto.request.CreatePostRequest;
+import ru.blog_app.blog.dto.response.PostDtoResponse;
 import ru.blog_app.blog.exception.PostNotFoundException;
 import ru.blog_app.blog.models.Post;
 import ru.blog_app.blog.models.User;
@@ -21,15 +23,15 @@ public class PostServiceImpl implements PostService {
     private final UserRepository userRepository;
     
     @Override
-    public PostDtoResponse createPost(PostDtoResponse postDtoResponse) {
-        User user = userRepository.findById(postDtoResponse.getUserId())
+    public PostDtoResponse createPost(CreatePostRequest request, String currentUsername) {
+        User user = userRepository.findByUsername(currentUsername)
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
         Post post = Post.builder()
-                .title(postDtoResponse.getTitle())
-                .content(postDtoResponse.getContent())
+                .title(request.getTitle())
+                .content(request.getContent())
                 .user(user)
-                .createTime(LocalDateTime.now())
+                .createdAt(LocalDateTime.now())
                 .build();
 
         Post savedPost = postRepository.save(post);
@@ -37,19 +39,28 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public void deletePost(Long id) {
+    public void deletePost(Long id, String currentUsername) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("Post not found"));
+
+        if (!post.getUser().getUsername().equals(currentUsername)) {
+            throw new AccessDeniedException("You can't delete this post");
+        }
+
         postRepository.delete(post);
     }
 
     @Override
-    public PostDtoResponse updatePost(PostDtoResponse postDtoResponse, Long id) {
+    public PostDtoResponse updatePost(CreatePostRequest request, Long id, String currentUsername) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new PostNotFoundException("Post not found"));
 
-        post.setTitle(postDtoResponse.getTitle());
-        post.setContent(postDtoResponse.getContent());
+        if (!post.getUser().getUsername().equals(currentUsername)) {
+            throw new AccessDeniedException("You can't update this post");
+        }
+
+        post.setTitle(request.getTitle());
+        post.setContent(request.getContent());
         Post updatedPost = postRepository.save(post);
 
         return convertToDto(updatedPost);
@@ -76,7 +87,7 @@ public class PostServiceImpl implements PostService {
                 .title(post.getTitle())
                 .content(post.getContent())
                 .userId(post.getUser().getId())
-                .createTime(post.getCreateTime())
+                .createTime(post.getCreatedAt())
                 .build();
     }
 }
